customer_service/requirements.txt
Django>=5.0,<7.0
djangorestframework>=3.15,<4.0
pymysql>=1.1,<2.0
requests>=2.31,<3.0
cryptography